+++ services/sapdiapi-service/Omicron.SapDiApi.Entities/Models/CompleteDetailModel.cs
@@ -35,7 +35,7 @@ namespace Omicron.SapDiApi.Entities.Models
         /// Gets or sets Code.
         /// </summary>
         /// <value>The code.</value>
-        public int? QtyPlanned { get; set; }
+        public decimal? QtyPlanned { get; set; }
 
         /// <summary>
         /// Gets or sets Code.
